feat(ui): migrate Icon to src/components/ui/
