+++ src/components/ui/Icon/index.ts
@@ -1,0 +1,3 @@
+export { Icon } from "./Icon";
+export type { IconProps } from "./Icon";
+export { default } from "./Icon";
